requirements.txt
keybert==0.8.5
gigachat==0.1.35
langchain_community==0.3.14
langdetect==1.0.9
numpy==1.26.4
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.41.1
tqdm==4.66.5
faiss-cpu==1.8.0.post1
